changing file structure
