+++ sktime/annotation/hmm_learn/__init__.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+# !/usr/bin/env python3 -u
+# copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
+"""Implements wrappers for estimators from hmmlearn."""
+
+__all__ = [
+    "BaseHMMLearn",
+    "GaussianHMM",
+    "GMMHMM",
+]
+
+from sktime.annotation.hmm_learn.base_HMM import BaseHMMLearn
+from sktime.annotation.hmm_learn.GaussianHMM import GaussianHMM
+from sktime.annotation.hmm_learn.GMMHMM import GMMHMM
